--- Sources/CBcrypt/include/module.modulemap
@@ -1,4 +1,0 @@
-module CBcrypt [system][extern_c] {
- header "../bcrypt.h"
- export *
-}
